pyproject.toml
[project]
name = "post_training_pipeline"
version = "0.1.0"
description = "LLM alignment pipeline: SFT, reward modeling, and DPO"
readme = "README.md"
requires-python = ">=3.11"
keywords = ["llm", "alignment", "sft", "dpo", "rlhf", "fine-tuning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "python-dotenv>=1.0.0",
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "peft>=0.10.0",
    "datasets>=2.18.0",
    "hydra-core>=1.3.0",
    "wandb>=0.16.0",
    "accelerate>=0.28.0",
    "bitsandbytes>=0.43.0",
    "trl>=0.8.0",
    "evaluate>=0.4.0",
    "omegaconf>=2.3.0",
    "safetensors>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.3.0",
]

[project.scripts]
post-train = "post_training_pipeline.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["post_training_pipeline"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501", "E402", "UP045"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
"scripts/*" = ["E402"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.3.0",
]
